# docs/openapi.yaml
openapi: 3.1.0
info:
  title: PairEval API (Campus Pairwise Evaluation Service)
  description: >
    RESTful API specification for **PairEval**, a campus pairwise peer and group evaluation platform.
    Designed according to PRD specifications, Architectural Principles (AR-01..06), Database Rules (DR-01..10),
    and all 31 User Stories (US-AUTH-01..04, US-CLASS-01..03, US-ASSIGN-01..03, US-PAIR-01..04, US-EVAL-01..05,
    US-SCORE-01..04, US-REPORT-01..03, US-EXPORT-01, US-INTEG-01, US-NOTIF-01, US-AUDIT-01, US-APPEAL-01).
  version: 1.0.0
  contact:
    name: PairEval Core Engineering Team
    email: support@paireval.university.ac.th
  license:
    name: Proprietary / University Internal
servers:
  - url: https://api.paireval.university.ac.th
    description: Production API Server
  - url: http://localhost:8000
    description: Local Development API Server
security:
  - BearerAuth: []
...
paths:
...
  /api/classrooms:
    get:
      tags:
        - Classrooms & Roster
      summary: List Classrooms for Current User
      description: >
        Lists all classrooms in which the authenticated user is enrolled as an Owner,
        Co-Teacher, TA, or Student (US-CLASS-01, US-AUTH-04).
      operationId: listClassrooms
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [ACTIVE, ARCHIVED, ALL]
            default: ACTIVE
          description: Filter by classroom operational status.
      responses:
        "200":
          description: List of classrooms.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClassroomSummary"
        "401":
          $ref: "#/components/responses/UnauthorizedError"

    post:
      tags:
        - Classrooms & Roster
      summary: Create a New Classroom
      description: >
        Creates a new classroom with a unique URL slug, timezone, and allowed email domains.
        The authenticated user is automatically assigned as the classroom Owner (US-CLASS-01).
      operationId: createClassroom
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClassroomRequest"
      responses:
        "201":
          description: Classroom successfully created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClassroomDetail"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "409":
          description: Classroom slug already exists in the system.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          $ref: "#/components/responses/UnprocessableEntityError"
  /api/classrooms/{classroomId}:
    get:
      tags:
        - Classrooms & Roster
      summary: Get Classroom Details
      description: >
        Retrieves detailed information for a specific classroom. Enforces classroom resource
        isolation (FR-AUTHZ-02 / US-AUTH-04), returning 404 if the user is not a member of this classroom.
      operationId: getClassroom
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ClassroomIdParam"
      responses:
        "200":
          description: Classroom details.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClassroomDetail"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"

    patch:
      tags:
        - Classrooms & Roster
      summary: Update or Archive Classroom
      description: >
        Updates classroom settings or changes status to ARCHIVED (read-only mode).
        Requires Owner or Co-Teacher role (US-CLASS-01).
      operationId: updateClassroom
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ClassroomIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClassroomRequest"
      responses:
        "200":
          description: Classroom successfully updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClassroomDetail"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "409":
          $ref: "#/components/responses/ConflictError"
        "422":
          $ref: "#/components/responses/UnprocessableEntityError"
...
  /api/classrooms/{classroomId}/roster:import:
    post:
      tags:
        - Classrooms & Roster
      summary: Atomic CSV Roster Import and Upsert
      description: >
        Imports or updates student roster and group allocations from a CSV file (`email`, `group_name`,
        `student_id`, `display_name`). Enforces atomic validation: fails the entire file if any row has an
        invalid email or formula injection character (`=`, `+`, `-`, `@`) (FR-SEC-04, US-CLASS-02).
        Supports `mode=upsert` with `dry_run=true` to preview diffs before committing (US-CLASS-03).
      operationId: importRosterCsv
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ClassroomIdParam"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: CSV file containing roster data (UTF-8 encoded).
                mode:
                  type: string
                  enum: [REPLACE, UPSERT]
                  default: UPSERT
                  description: REPLACE resets existing roster; UPSERT updates student groups.
                dry_run:
                  type: boolean
                  default: false
                  description: If true, returns diff preview without saving changes.
      responses:
        "200":
          description: Dry run diff preview successfully generated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RosterImportResult"
        "201":
          description: Roster CSV parsed and imported successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RosterImportResult"
        "400":
          description: CSV parsing or row validation error (includes specific row number and reason).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "422":
          description: Validation error such as group size < 2 or formula injection detected.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/classrooms/{classroomId}/members:
    post:
      tags:
        - Classrooms & Roster
      summary: Add Co-Teacher or TA to Classroom
      description: >
        Adds a Co-Teacher or TA to the classroom by normalized university email (US-CLASS-03).
      operationId: addClassroomMember
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ClassroomIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddMemberRequest"
      responses:
        "201":
          description: Classroom member successfully added.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClassroomMember"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "409":
          description: User is already a member of this classroom.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          $ref: "#/components/responses/UnprocessableEntityError"
  /api/classrooms/{classroomId}/members/{memberId}:
    delete:
      tags:
        - Classrooms & Roster
      summary: Remove Member from Classroom
      description: >
        Removes a Co-teacher, TA, or Student from the classroom. Enforces safety invariant:
        the last remaining Owner cannot be removed (US-CLASS-03, FR-CLASS-06).
      operationId: removeClassroomMember
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ClassroomIdParam"
        - name: memberId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Unique identifier of the classroom member record.
      responses:
        "204":
          description: Member successfully removed.
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "409":
          description: Cannot remove the last remaining Owner from the classroom.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ==========================================
  # 3. Assignments & Criteria Setup
  # ==========================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Google OIDC verified JWT session token. Must be passed in the `Authorization: Bearer <token>` header.
  parameters:
    ClassroomIdParam:
      name: classroomId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Unique identifier (UUID) of the classroom.
...
  responses:
    BadRequestError:
      description: Bad Request — Request payload is malformed, invalid JSON, or unparseable.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: BAD_REQUEST
              message: Invalid JSON syntax or request parameters.
              field: null
              requestId: 01J9X8K9Z2M4NP7Q3R1T5V8W0Y
    UnauthorizedError:
      description: Unauthorized — Missing, expired, or invalid Bearer authentication token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: UNAUTHORIZED
              message: Missing or invalid authentication credentials.
              field: null
              requestId: 01J9X8K9Z2M4NP7Q3R1T5V8W0Y
    ForbiddenError:
      description: Forbidden — Authenticated user lacks the necessary role permissions (FR-AUTHZ-01).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: FORBIDDEN
              message: Access forbidden. Student role cannot access instructor reports.
              field: null
              requestId: 01J9X8K9Z2M4NP7Q3R1T5V8W0Y
    NotFoundError:
      description: Not Found — Resource does not exist or belongs to another classroom (FR-AUTHZ-02).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: NOT_FOUND
              message: The requested resource was not found.
              field: null
              requestId: 01J9X8K9Z2M4NP7Q3R1T5V8W0Y
    ConflictError:
      description: Conflict — Action violates state machine transition or uniqueness constraint.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: STATE_CONFLICT
              message: Cannot modify criteria when assignment is not in DRAFT status.
              field: status
              requestId: 01J9X8K9Z2M4NP7Q3R1T5V8W0Y
    UnprocessableEntityError:
      description: Unprocessable Entity — Schema validation failed (e.g. weights != 100%, choice != 1..6).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: VALIDATION_ERROR
              message: Criteria weight sum must equal 100.00% (currently 95.00%).
              field: criteria
              requestId: 01J9X8K9Z2M4NP7Q3R1T5V8W0Y
...
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          $ref: "#/components/schemas/ErrorDetail"
    ErrorDetail:
      type: object
      required:
        - code
        - message
        - requestId
      properties:
        code:
          type: string
          description: Standardized machine-readable error code.
          example: DEADLINE_PASSED
        message:
          type: string
          description: Human-readable error description in Thai or English.
          example: The evaluation deadline for this assignment has passed.
        field:
          type: string
          nullable: true
          description: Specific field that caused the error, if applicable.
          example: individual_deadline_utc
        requestId:
          type: string
          description: Unique request correlation ID for tracing.
          example: 01J9X8K9Z2M4NP7Q3R1T5V8W0Y
        details:
          type: array
          description: Granular validation failure items (e.g. CSV row errors).
          items:
            $ref: "#/components/schemas/ErrorRowDetail"
    ErrorRowDetail:
      type: object
      required:
        - row
        - message
      properties:
        row:
          type: integer
          description: CSV row index where validation failed.
          example: 42
        field:
          type: string
          description: Field name in the CSV row.
          example: email
        message:
          type: string
          description: Specific error detail for this row.
          example: Invalid email format

    # ----------------------------------------
    # Auth Schemas
    # ----------------------------------------
...
    ClassroomSummary:
      type: object
      required:
        - id
        - name
        - slug
        - timezone
        - status
        - my_role
        - created_at
      properties:
        id:
          type: string
          format: uuid
          example: 11111111-2222-3333-4444-555555555555
        name:
          type: string
          example: Software Engineering 2026/1
        slug:
          type: string
          example: cpe-se-2026
        timezone:
          type: string
          example: Asia/Bangkok
        status:
          type: string
          enum: [ACTIVE, ARCHIVED]
          example: ACTIVE
        my_role:
          type: string
          enum: [OWNER, CO_TEACHER, TA, STUDENT]
          example: OWNER
        created_at:
          type: string
          format: date-time
          example: 2026-08-01T09:00:00Z
    CreateClassroomRequest:
      type: object
      required:
        - name
        - slug
        - timezone
        - allowed_email_domains
      properties:
        name:
          type: string
          example: Software Engineering 2026/1
        slug:
          type: string
          example: cpe-se-2026
        timezone:
          type: string
          example: Asia/Bangkok
        allowed_email_domains:
          type: array
          items:
            type: string
          example:
            - uni.ac.th
            - student.uni.ac.th
    UpdateClassroomRequest:
      type: object
      properties:
        name:
          type: string
          example: Software Engineering 2026/1 (Updated)
        timezone:
          type: string
          example: Asia/Bangkok
        allowed_email_domains:
          type: array
          items:
            type: string
          example:
            - uni.ac.th
        status:
          type: string
          enum: [ACTIVE, ARCHIVED]
          example: ARCHIVED
    ClassroomDetail:
      type: object
      required:
        - id
        - name
        - slug
        - timezone
        - allowed_email_domains
        - status
        - created_by
        - created_at
      properties:
        id:
          type: string
          format: uuid
          example: 11111111-2222-3333-4444-555555555555
        name:
          type: string
          example: Software Engineering 2026/1
        slug:
          type: string
          example: cpe-se-2026
        timezone:
          type: string
          example: Asia/Bangkok
        allowed_email_domains:
          type: array
          items:
            type: string
          example:
            - uni.ac.th
        status:
          type: string
          enum: [ACTIVE, ARCHIVED]
          example: ACTIVE
        created_by:
          type: string
          format: uuid
          example: a1b2c3d4-e5f6-7a8b-9c0d-1e2f3a4b5c6d
        created_at:
          type: string
          format: date-time
          example: 2026-08-01T09:00:00Z
    ClassroomMember:
      type: object
      required:
        - id
        - user_id
        - email
        - display_name
        - role
        - joined_at
      properties:
        id:
          type: string
          format: uuid
          example: 33333333-4444-5555-6666-777777777777
        user_id:
          type: string
          format: uuid
          example: a1b2c3d4-e5f6-7a8b-9c0d-1e2f3a4b5c6d
        email:
          type: string
          format: email
          example: somchai.a@uni.ac.th
        display_name:
          type: string
          example: Somchai Anan
        role:
          type: string
          enum: [OWNER, CO_TEACHER, TA, STUDENT]
          example: STUDENT
        group_id:
          type: string
          format: uuid
          nullable: true
          example: 22222222-3333-4444-5555-666666666666
        group_name:
          type: string
          nullable: true
          example: Aurora
        joined_at:
          type: string
          format: date-time
          example: 2026-08-01T10:00:00Z
    AddMemberRequest:
      type: object
      required:
        - email
        - role
      properties:
        email:
          type: string
          format: email
          example: co_teacher@uni.ac.th
        role:
          type: string
          enum: [CO_TEACHER, TA]
          example: CO_TEACHER
...
    RosterImportResult:
      type: object
      required:
        - mode
        - dry_run
        - total_rows
        - students_created
        - students_updated
        - groups_created
        - groups_updated
        - warnings
      properties:
        mode:
          type: string
          enum: [REPLACE, UPSERT]
          example: UPSERT
        dry_run:
          type: boolean
          example: false
        total_rows:
          type: integer
          example: 45
        students_created:
          type: integer
          example: 40
        students_updated:
          type: integer
          example: 5
        groups_created:
          type: integer
          example: 9
        groups_updated:
          type: integer
          example: 0
        diff:
          type: array
          description: Diff entries when dry_run=true or upserting.
          items:
            type: object
            required:
              - email
              - action
            properties:
              email:
                type: string
                example: somchai.a@uni.ac.th
              action:
                type: string
                enum: [ADDED, MOVED_GROUP, UNCHANGED]
                example: MOVED_GROUP
              old_group:
                type: string
                nullable: true
                example: Alpha
              new_group:
                type: string
                example: Aurora
        warnings:
          type: array
          items:
            type: string
          example:
            - 'Group "Zeta" has only 2 members; recommended size is 4-7 students.'

    # ----------------------------------------
    # Assignment & Criteria Schemas
    # ----------------------------------------
...